Home added and lots more. - Added dashboard into the Home screen. - Enhanced the light mode by changing the background colour to a bit darker. - Added Toast message when Update/Create. - Added invalid tooltip when invalid value is inputted. - Added Settings > Category - Added "Settings > Delete All Data" for testing purpose.

+++ FinPal/Models/Salary.cs
@@ -12,7 +12,7 @@ namespace FinPal.Models
     {
         [PrimaryKey, AutoIncrement]
         public int Id { set; get; }
-        public int Amount { set; get; }
+        public decimal Amount { set; get; }
         public int Month { set; get; }
         public int Year { set; get; }
         public bool Active { set; get; }
